Renamed templates_apply.py and separated parse_logcluster into another file.
